old/requirements_and_mkdocs.txt
altgraph==0.17.4
babel==2.17.0
backrefs==5.8
certifi==2025.1.31
charset-normalizer==3.4.1
chess==1.11.2
click==8.1.8
colorama==0.4.6
ghp-import==2.1.0
idna==3.10
Jinja2==3.1.6
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.11
mkdocs-material-extensions==1.3.1
mkdocstrings==0.29.1
packaging==24.2
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.7
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pymdown-extensions==10.14.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==0.1
requests==2.32.3
setuptools==77.0.1
six==1.17.0
urllib3==2.3.0
watchdog==6.0.0
